% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{pkgsync}
\alias{pkgsync}
\alias{pacman_ls}
\alias{pacman}
\alias{pacman_download}
\title{Download Full Pacman Repository}
\usage{
pacman_ls(url, repo = basename(url))

pacman_download(url, repo = basename(url), destdir = basename(url))
}
\arguments{
\item{url}{full URL to the location of the repo}

\item{repo}{name of the repository db files. Usually this is equal
to the dirname.}

\item{destdir}{where to store downloaded files.}
}
\description{
List and download all packages in a pacman repository. Useful for
mirroring the repository locally or to another server.
}
\examples{
\donttest{
pacman_download('http://dl.bintray.com/rtools/mingw32/')
pacman_download('http://dl.bintray.com/rtools/mingw64/')
}
}
